backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
google-cloud-vision>=3.7.0
google-cloud-speech>=2.26.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
anthropic>=0.40.0
python-multipart>=0.0.9
httpx>=0.27.0
weasyprint>=61.0
Jinja2>=3.1.4
Pillow>=10.3.0
python-jose[cryptography]>=3.3.0
openai>=1.30.0
reportlab>=4.2.0
aiofiles>=23.2.1
python-dotenv>=1.0.1
cachetools>=5.3.3
